general cleanup; move functions to utilities module
