+++ src/vivarium_nih_us_cvd/utilities.py
@@ -9,7 +9,7 @@ from scipy import stats
 from vivarium.framework.randomness import get_hash
 from vivarium_public_health.risks.data_transformations import pivot_categorical
 
-from vivarium_nih_us_cvd.constants import metadata
+from vivarium_nih_us_cvd.constants import data_values, metadata
 
 SeededDistribution = Tuple[str, stats.rv_continuous]
 
@@ -176,3 +176,33 @@ def get_random_variable(draw: int, seeded_distribution: SeededDistribution) -> f
     seed, distribution = seeded_distribution
     np.random.seed(get_hash(f"{seed}_draw_{draw}"))
     return distribution.rvs()
+
+def get_measurement_error(index: pd.Index, mean: float, sd: float, randomness: "RandomnessStream") -> pd.Series:
+        """Return measurement error assuming normal distribution"""
+        draw = randomness.get_draw(index)
+        return stats.norm(loc=mean, scale=sd).ppf(draw)
+
+def schedule_followup(
+    index: pd.Index,
+    event_time: pd.Timestamp,
+    randomness: "RandomnessStream",
+    min_followup: int = data_values.FOLLOWUP_MIN,
+    max_followup: int = data_values.FOLLOWUP_MAX,
+) -> pd.Series:
+    """Schedules followup visits"""
+    return pd.Series(
+        event_time
+        + random_time_delta(
+            pd.Series(min_followup, index=index),
+            pd.Series(max_followup, index=index),
+            randomness,
+        ),
+        index=index,
+    )
+
+def random_time_delta(start: pd.Series, end: pd.Series, randomness: "RandomnessStream") -> pd.Series:
+    """Generate a random time delta for each individual in the start
+    and end series."""
+    return pd.to_timedelta(
+        start + (end - start) * randomness.get_draw(start.index), unit="day"
+    )
